djangoapp/requirements.txt
Django
djangorestframework
numpy
scikit_learn
tensorflow
opencv-python